db fix

+++ internal/migrations/20240829113459_create_users_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE public.users (
                                  id uuid DEFAULT gen_random_uuid() NOT NULL,
                                  login varchar(255) NOT NULL,
                                  password varchar(60) NOT NULL,
-                                 balance integer DEFAULT 0 NOT NULL CHECK (balance >= 0),
+                                 balance numeric(12, 2) DEFAULT 0 NOT NULL CHECK (balance >= 0),
                                  created_at timestamptz DEFAULT now() NOT NULL,
                                  CONSTRAINT users_pkey PRIMARY KEY (id)
 );
